requirements.txt
# requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.0
matplotlib>=3.6.0
python-dotenv>=1.0.0
joblib>=1.2.0
tqdm>=4.64.0
pytest>=7.0.0
flask>=2.2.0  # Optional for API